+++ packages/af-utils-react-virtual-headless/src/hooks/useVirtual/index.js
@@ -1,10 +1,12 @@
-import { useLayoutEffect } from "react";
+import { useLayoutEffect, useEffect } from "react";
 import useVirtualModel from "../useVirtualModel";
+
+const localUseEffect = process.env.__IS_SERVER__ ? useEffect : useLayoutEffect;
 
 const useVirtual = params => {
     const model = useVirtualModel(params);
 
-    useLayoutEffect(() => {
+    localUseEffect(() => {
         model._startBatch();
         model.setHorizontal(params.horizontal);
         model.setItemCount(params.itemCount, params.getEstimatedItemSize);