+++ packages/website/components/MdxWrapper/index.js
@@ -1,6 +1,6 @@
 const MdxWrapper = ({ children }) => (
     <div className="overflow-auto w-full px-6">
-        <div className="prose md:prose-lg max-w-[60em] pt-4">{children}</div>
+        <div className="prose prose-neutral max-w-[80em] pt-4">{children}</div>
     </div>
 );