+++ packages/af-utils-react-mobx-table/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/react-mobx-table
 
+## 0.0.22
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.16
+    -   @af-utils/react-table@0.0.20
+
 ## 0.0.21
 
 ### Patch Changes